requirements.txt
streamlit
python-chess
pyyaml
cairosvg
psutil
pillow
pandas
torch
numpy
dash
plotly